requirements.txt
alpaca_trade_api
python_dotenv
ta-lib
pandas-market-calendars
pandas
numpy